support numeric enum property values >127 (beyond Java byte), define initial Script property values
git-svn-id: https://unicode.org/repos/unicodetools/trunk@742 13e8329f-0b23-4da4-9fe8-d0f6fe080806

+++ unicodetools/org/unicode/text/UCA/Fractional.java
@@ -264,8 +264,8 @@ public final class Fractional {
         return temp.toString();
     }
 
-    /* package */ static byte getFixedScript(int ch) {
-        byte script = Default.ucd().getScript(ch);
+    /* package */ static short getFixedScript(int ch) {
+        short script = Default.ucd().getScript(ch);
         // HACK
         if (ch == 0x0F7E || ch == 0x0F7F) {
             if (script != UCD_Types.TIBETAN_SCRIPT) {
